requirements.txt
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.8.0
argcomplete==3.5.3
asttokens==3.0.0
attrs==25.1.0
azure-ai-evaluation==1.2.0
azure-core==1.32.0
azure-identity==1.20.0
azure-monitor-opentelemetry-exporter==1.0.0b33
azure-storage-blob==12.24.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
cryptography==44.0.1
debugpy==1.8.12
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
docstring_parser==0.16
executing==2.2.0
fastapi==0.115.8
filelock==3.17.0
filetype==1.2.0
fixedint==0.1.6
Flask==3.1.0
flask-cors==5.0.1
flask-restx==1.3.0
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.68.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.32.0
isodate==0.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==24.3.1
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
more-itertools==10.6.0
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.2.3
oauthlib==3.2.2
openai==1.64.0
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
portalocker==2.10.1
prompt_toolkit==3.0.50
promptflow-core==1.17.2
promptflow-devkit==1.17.2
promptflow-tracing==1.17.2
protobuf==5.29.3
psutil==5.9.8
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydash==7.0.7
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==308
pywin32-ctypes==0.2.3
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.45.3
strictyaml==1.7.3
tabulate==0.9.0
tiktoken==0.9.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
waitress==3.0.2
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
